refactor(banner): move banner to communication app
