+++ app/communication/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from app.communication.models import Mail, Notification, Warning
+from app.communication.models import Mail, Notification, Warning, Banner
 
 
 class MailAdmin(admin.ModelAdmin):
@@ -14,3 +14,4 @@ class MailAdmin(admin.ModelAdmin):
 admin.site.register(Mail, MailAdmin)
 admin.site.register(Notification)
 admin.site.register(Warning)
+admin.site.register(Banner)
