+++ app/communication/factories/__init__.py
@@ -1,2 +1,3 @@
+from app.communication.factories.banner_factory import BannerFactory
 from app.communication.factories.mail_factory import MailFactory
 from app.communication.factories.notification_factory import NotificationFactory
