+++ app/communication/models/__init__.py
@@ -1,3 +1,4 @@
+from app.communication.models.banner import Banner
 from app.communication.models.mail import Mail
 from app.communication.models.notification import Notification
 from app.communication.models.warning import Warning
