+++ app/communication/serializers/__init__.py
@@ -1,3 +1,4 @@
+from app.communication.serializers.banner import BannerSerializer
 from app.communication.serializers.notification import (
     NotificationSerializer,
     UpdateNotificationSerializer,
