+++ app/communication/views/__init__.py
@@ -1,2 +1,3 @@
+from app.communication.views.banner import BannerViewSet
 from app.communication.views.notification import NotificationViewSet
 from app.communication.views.warning import WarningViewSet
