+++ app/communication/views/banner.py
@@ -2,8 +2,8 @@ from rest_framework import viewsets
 
 from app.common.pagination import BasePagination
 from app.common.permissions import BasicViewPermission
-from app.content.models.banner import Banner
-from app.content.serializers.banner import BannerSerializer
+from app.communication.models.banner import Banner
+from app.communication.serializers.banner import BannerSerializer
 
 
 class BannerViewSet(viewsets.ModelViewSet):
