+++ app/content/models/__init__.py
@@ -1,5 +1,4 @@
 from app.content.models.badge import Badge
-from app.content.models.banner import Banner
 from app.content.models.category import Category
 from app.content.models.cheatsheet import Cheatsheet
 from app.content.models.event import Event
