+++ app/content/serializers/__init__.py
@@ -1,5 +1,4 @@
 from app.content.serializers.badge import BadgeSerializer
-from app.content.serializers.banner import BannerSerializer
 from app.content.serializers.category import CategorySerializer
 from app.content.serializers.cheatsheet import CheatsheetSerializer
 from app.content.serializers.event import (
