+++ app/content/urls.py
@@ -3,7 +3,6 @@ from django.urls import path
 from rest_framework import routers
 
 from app.content.views import (
-    BannerViewSet,
     CategoryViewSet,
     CheatsheetViewSet,
     EventViewSet,
@@ -38,7 +37,6 @@ router.register(
 router.register("badges", UserBadgeViewSet, basename="badge")
 router.register("pages", PageViewSet)
 router.register("strikes", StrikeViewSet, basename="strikes")
-router.register("banners", BannerViewSet, basename="banner")
 
 urlpatterns = [
     url(r"", include(router.urls)),
