+++ app/content/views/__init__.py
@@ -1,6 +1,5 @@
 from app.content.views.category import CategoryViewSet
 from app.content.views.cheatsheet import CheatsheetViewSet
-from app.content.views.banner import BannerViewSet
 from app.content.views.event import EventViewSet
 from app.content.views.user import UserViewSet
 from app.content.views.user_badge import UserBadgeViewSet
