+++ app/tests/communication/test_banner_integration.py
